% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanupForMarkdown_univAndMultiGLM.R
\name{cleanupForMarkdown_univAndMultiGLM}
\alias{cleanupForMarkdown_univAndMultiGLM}
\title{Extract output from univAndMultiGLM}
\usage{
cleanupForMarkdown_univAndMultiGLM(mod_matrix)
}
\arguments{
\item{mod_matrix}{Output from univAndMultiGLM}
}
\description{
This is a wrapper around univAndMultiGLM to print correctly in Rmd files
}
\examples{
cleanupForMarkdown_univAndMultiGLM()



}
