% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COCreateTableOne.r
\name{COCreateTableOne}
\alias{COCreateTableOne}
\title{Wrapper around tableone to print in RMD}
\usage{
COCreateTableOne(data_here, ..., caption = "Table", catDigits = 1,
  contDigits = 2, pDigits = 3, missing = FALSE, explain = TRUE,
  test = TRUE, smd = FALSE, format = c("fp", "f", "p", "pf")[1],
  showAllLevels = FALSE, cramVars = NULL, dropEqual = FALSE,
  exact = NULL, nonnormal = NULL, minMax = FALSE, full_width = F)
}
\arguments{
\item{data_here}{dataset to summarize all columns}

\item{caption}{able title}

\item{strata}{string to group by, NULL default}
}
\description{
This function allows you format the tidy output from broom in a more readable way
}
\keyword{tableone}
