% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mod_zeroinf_count.R
\name{extract_mod_zeroinf_count}
\alias{extract_mod_zeroinf_count}
\title{Extract nice coef table for negative binomial model}
\usage{
extract_mod_zeroinf_count(mod_zeroinf, digits = 2)
}
\arguments{
\item{mod_matrix}{Output from mod_zeroinf}
}
\description{
This is a wrapper around univAndMultiGLM which presents both binomial and Poisson coefficients
}
\examples{
\dontrun{
extract_mod_zeroinf_count()
}

}
