% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_crr.R
\name{format_crr}
\alias{format_crr}
\title{Print competing risks regression output with CI and P values}
\usage{
format_crr(crr_mod, d = 2, d_pVal = 3)
}
\arguments{
\item{crr_mod}{Output from cmprsk}

\item{d}{Decimal places used for HR and confidence interval}

\item{d_pVal}{Decimal places used for p value.}
}
\description{
This is a wrapper around broom:::tidy that adds formatted confidence intervals and p values
}
\examples{
\dontrun{
format_crr()
}

}
