% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pval.R
\name{format_pval}
\alias{format_pval}
\title{Format p value}
\usage{
format_pval(raw_p, d = 3, equal = "=")
}
\arguments{
\item{raw_p}{Number to be treated as p value}

\item{d}{Decimal places used for p value.}

\item{equal}{Symbol to include if not <}
}
\description{
This function allows you format the o values including <.001
}
\examples{
format_pval(0.0001)
format_pval(0.95,equal = "")

}
