% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tidy_cox.R
\name{format_tidy_cox}
\alias{format_tidy_cox}
\title{Print coxph after using tidy}
\usage{
format_tidy_cox(rslt, d = 2, d_pVal = 3)
}
\arguments{
\item{rslt}{Output from tidy(coxph(),exponentiate=T)}

\item{d}{Decimal places used for HR and confidence interval.}

\item{d_pVal}{Decimal places used for p value.}
}
\description{
This function allows you format the tidy output from broom in a more readable way
}
\examples{
\dontrun{
#format_tidy_cox()
}
}
\keyword{tidy}
