% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tidy_coxme.R
\name{format_tidy_coxme}
\alias{format_tidy_coxme}
\title{Print Cox Mixed effects regression output with CI and P values}
\usage{
format_tidy_coxme(mod, d = 2, d_pVal = 3, sub_out = NULL, sub_in = NULL)
}
\arguments{
\item{d}{Decimal places used for HR and confidence interval}

\item{d_pVal}{Decimal places used for p value.}

\item{sub_out}{String to from variable names to replace with sub_in}

\item{sub_in}{String to from variable names to replace instead of sub_out}

\item{crr_mod}{Output from coxme}
}
\description{
This is a wrapper around broom:::tidy that adds formatted confidence intervals and p values
}
\examples{
format_tidy_coxme()
}
