% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tidy_gen.R
\name{format_tidy_gen}
\alias{format_tidy_gen}
\title{Print generic regression output with CI and P values}
\usage{
format_tidy_gen(rslt, d = 2, d_pVal = 3, sub_out = NULL, sub_in = NULL,
  exponentiate_here = FALSE, Beta = "Beta", reportIntercept = FALSE)
}
\arguments{
\item{rslt}{Output from regression (lm, glm, or coxPH)}

\item{d}{Decimal places used for HR and confidence interval}

\item{d_pVal}{Decimal places used for p value.}

\item{sub_out}{String to from variable names to replace with sub_in}

\item{sub_in}{String to from variable names to replace instead of sub_out}

\item{exponentiate_here}{Select whether exponentiation should be done (CoxPH, Poisson, etc.). Default is FALSE}

\item{Beta}{Label for coefficient. Default is 'Beta'}

\item{reportIntercept}{Whether intercept should be reported. Default is FALSE.}
}
\description{
This is a wrapper around broom:::tidy that adds formatted confidence intervals and p values
}
\examples{
dontrun{
format_tidy_gen()
}
}
\keyword{tidy}
