% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tidy_lm.R
\name{format_tidy_lm}
\alias{format_tidy_lm}
\title{Print linear model after using tidy}
\usage{
format_tidy_lm(rslt, d = 2, d_pVal = 3)
}
\arguments{
\item{rslt}{Output from lm() then tidy()}

\item{d}{Decimal places used for HR and confidence interval}

\item{d_pVal}{Decimal places used for p value.}
}
\description{
This function allows you format the tidy output from broom in a more readable way. This has been deprecated and replaced with format_tidy_gen, which also has more options.
}
\examples{
dontrun{
format_tidy_lm()
}

}
\keyword{tidy}
