% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktab_ChiSq_postHoc.R
\name{mktab_ChiSq_postHoc}
\alias{mktab_ChiSq_postHoc}
\title{Wrapper around tableone to print in RMD}
\usage{
mktab_ChiSq_postHoc(var, grp, varNm = "Var", file = "", notes = "",
  size = "scriptsize", d = 3, control = "bonferroni", markdown = FALSE)
}
\arguments{
\item{var}{vector of counts}

\item{grp}{vector of groupings of interest}

\item{varNm}{label for caption}

\item{control}{type of correction}
}
\description{
This function exports results of Chi-sq Post Hoc Test in a legible manner
}
\examples{
\dontrun{
mktab_ChiSq_postHoc(mtcars)
}
}
\keyword{tableone}
