% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktab_KW_postHoc.R
\name{mktab_KW_postHoc}
\alias{mktab_KW_postHoc}
\title{Print results of Kruskal-Wallis Post Hoc Test}
\usage{
mktab_KW_postHoc(var, grp, varNm = "Var", file = "", notes = "",
  size = "scriptsize", d = 3)
}
\arguments{
\item{var}{vector of counts}

\item{grp}{vector of groupings of interest}

\item{varNm}{label for caption}
}
\description{
This function exports results of KW post hoc test in a legible manner
}
\examples{
\dontrun{
mktab_KW_postHoc(mtcars)
}
}
\keyword{tableone}
