% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLinear_coefTable.R
\name{myLinear_coefTable}
\alias{myLinear_coefTable}
\title{Extract nice coef table for linear model}
\usage{
myLinear_coefTable(glm_linear)
}
\arguments{
\item{glm_linear}{Output from glm(, family = "gaussian")}
}
\description{
This is a wrapper around glm to print nicely. Could use broom, but keeping here for backwards compatibility.
}
\examples{
myLinear_coefTable()


}
