% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLogistic_coefTable.R
\name{myLogistic_coefTable}
\alias{myLogistic_coefTable}
\title{Extract nice coef table for  binomial model}
\usage{
myLogistic_coefTable(glm_log)
}
\arguments{
\item{glm_log}{Output from glm, family = "binomial"}
}
\description{
This is a wrapper around glm to print nicely. Could use broom, but keeping here for backwards compatibility.
}
\examples{
myLogistic_coefTable()


}
