% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univAndMultiGLM.R
\name{univAndMultiGLM}
\alias{univAndMultiGLM}
\title{Make a table with univariable and multivariable logistic results}
\usage{
univAndMultiGLM(outcome, covars_uni, covars_multi = NULL, p_multiV = 0.05,
  family = "binomial", file = "", d = 2, d_pVal = 3, CIs = TRUE,
  z0 = 1.96, name = "GLM", widthCharCol = 6, markdown = FALSE)
}
\arguments{
\item{outcome}{Vector of outcome}

\item{covars_uni}{Matrix with columns to be used in univariable models}

\item{covars_multi}{Matrix with columns to be used in multivariable models. If NULL, univariable predictors with p < 0.05 will be used.}

\item{p_multiV}{ANOVA p-value threshold used for inclusion in multiariable model. Ignored if covars_multi passed in.}

\item{family}{Model family for GLM: "binomial", "gaussian", "poisson".}

\item{file}{File name for LaTeX output to be sent to.}

\item{d}{Decimal places used for HR and confidence interval.}

\item{d_pVal}{Decimal places used for p value.}

\item{CIs}{Logical to indicate whether confidence intervals should be printed.}

\item{z0}{Critical value used for confidence interval.}

\item{name}{Name of model to be used for caption}

\item{widthCharCol}{Width of first column with predictor names and levels.}

\item{markdown}{Boolean to indicate that output be compatible with markdown rather than LaTeX.}
}
\description{
Given a dataset, make two column output with univariable results for each predictor, and second column with single multivariable model
}
\examples{
dontrun{
univAndMultiGLM()
}

}
